+++ Renderer/RendererBackend/OpenGLES2Renderer/src/Buffer/VertexArrayNoVao.cpp
@@ -125,7 +125,7 @@ namespace OpenGLES2Renderer
 
 		#ifndef OPENGLES2RENDERER_NO_STATE_CLEANUP
 			// Be polite and restore the previous bound OpenGL ES 2 array buffer
-			glBindBuffer(GL_ARRAY_BUFFER, openGLES2ArrayBufferBackup);
+			glBindBuffer(GL_ARRAY_BUFFER, static_cast<GLuint>(openGLES2ArrayBufferBackup));
 		#endif
 
 		// Get the used index buffer
